gprMax/requirements.txt
#absl-py==0.7.0
appdirs==1.4.3
astor==0.7.1
astroid==2.1.0
atomicwrites==1.2.1
attrs==18.2.0
autopep8==1.4.3
backcall==0.1.0
bleach==3.1.0
blessings==1.7
certifi==2018.11.29
chardet==3.0.4
#codecov==2.0.15
colorama==0.4.1
colour-runner==0.0.5
coverage==4.5.2
cycler==0.10.0
Cython==0.29.3
decorator==4.3.2
defusedxml==0.5.0
entrypoints==0.3
espatools==0.0.8
gast==0.2.2
gprMax==3.1.4
grpcio==1.18.0
h5py==2.9.0
idna==2.8
imageio==2.4.1
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.4
jedi==0.13.2
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.4.0
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.8
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
lxml==4.3.0
Mako==1.0.7
Markdown==3.0.1
MarkupSafe==1.1.0
matplotlib==3.0.2
mccabe==0.6.1
mistune==0.8.4
mock==2.0.0
more-itertools==5.0.0
mpi4py==3.0.0
nbconvert==5.4.0
nbformat==4.4.0
notebook==5.7.4
numpy==1.16.1
pandas==0.24.0
pandocfilters==1.4.2
parso==0.3.2
pbr==5.1.2
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
pluggy==0.8.1
prometheus-client==0.5.0
prompt-toolkit==2.0.8
properties==0.5.5
protobuf==3.6.1
psutil==5.5.0
ptyprocess==0.6.0
PVGeo==1.2.1
py==1.7.0
pycodestyle==2.5.0
pycuda==2018.1.1
Pygments==2.3.1
pylint==2.2.2
pyparsing==2.3.1
pytest==4.2.0
python-dateutil==2.7.5
pytools==2019.1
pytz==2018.9
PyYAML==3.13
pyzmq==17.1.2
qtconsole==4.4.3
requests==2.21.0
scikit-learn==0.20.2
scipy==1.2.0
Send2Trash==1.5.0
six==1.12.0
sklearn==0.0
stevedore==1.30.0
tensorboard==1.12.2
tensorflow-gpu==1.12.0
termcolor==1.1.0
terminado==0.8.1
terminaltables==3.1.0
testpath==0.4.2
tornado==5.1.1
tqdm==4.30.0
traitlets==4.3.2
typed-ast==1.3.0
urllib3==1.24.1
vectormath==0.2.1
virtualenv==16.3.0
virtualenv-clone==0.5.1
virtualenvwrapper==4.8.2
vtk==8.1.2
vtki==0.16.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
wrapt==1.11.1
xmltodict==0.11.0